WIP cleanning
