+++ lib/ui/instrumentpanel.js
@@ -1,7 +1,6 @@
 import BaconModel from 'bacon.model';
 import compareVersions from 'compare-versions';
 import Debug from 'debug';
-import isEqual from 'lodash/isEqual';
 
 const debug = Debug('instrumentpanel:instrumentpanel')
