+++ lib/ui/main.js
@@ -30,7 +30,7 @@ import {
   CS_DARK,
   backupSettingsKeyName
 } from './settings/constants';
-import WidthProvider from '../util/WidthProvider-new';
+import WidthProvider from '../util/WidthProvider';
 
 const debug = Debug('instrumentpanel:main')
