+++ lib/ui/navbar.js
@@ -28,8 +28,6 @@ import {
   CS_BY_MAINBAR
 } from './settings/constants'
 
-var buttonClass = "navbar-btn";
-
 export default class IpNavBar extends React.Component {
   constructor(props) {
     super(props);
