+++ lib/util/WidthProvider.jsx
@@ -1,32 +1,48 @@
-// from https://github.com/STRML/react-grid-layout/blob/master/lib/components/WidthProvider.jsx
 // @flow
-import React from "react";
+import * as React from "react";
 import PropTypes from "prop-types";
-import ReactDOM from "react-dom";
-import type { ComponentType as ReactComponentType } from "react";
+import clsx from "clsx";
+import type { ReactRef } from "../ReactGridLayoutPropTypes";
 
-type WPProps = {
+type WPDefaultProps = {|
+  measureBeforeMount: boolean
+    |};
+
+// eslint-disable-next-line no-unused-vars
+type WPProps = {|
   className?: string,
-  measureBeforeMount: boolean,
-  style?: Object,
-  reduceWidth: number
-};
+    style ?: Object,
+  ...WPDefaultProps
+  |};
 
 type WPState = {|
   width: number
-|};
+    |};
+
+type ComposedProps<Config> = {|
+  ...Config,
+  measureBeforeMount ?: boolean,
+  className ?: string,
+  style ?: Object,
+  width ?: number
+    |};
+
+const layoutClassName = "react-grid-layout";
 
 /*
  * A simple HOC that provides facility for listening to container resizes.
+ *
+ * The Flow type is pretty janky here. I can't just spread `WPProps` into this returned object - I wish I could - but it triggers
+ * a flow bug of some sort that causes it to stop typechecking.
  */
-export default function WidthProvider<
-  Props,
-  ComposedProps: { ...Props, ...WPProps }
->(
-  ComposedComponent: ReactComponentType<Props>
-): ReactComponentType<ComposedProps> {
-  return class WidthProvider extends React.Component<ComposedProps, WPState> {
-    static defaultProps = {
+export default function WidthProvideRGL<Config>(
+  ComposedComponent: React.AbstractComponent<Config>
+): React.AbstractComponent<ComposedProps<Config>> {
+  return class WidthProvider extends React.Component<
+    ComposedProps<Config>,
+    WPState
+  > {
+    static defaultProps: WPDefaultProps = {
       measureBeforeMount: false,
       reduceWidth: 0
     };
@@ -38,15 +54,15 @@ export default function WidthProvider<
       reduceWidth: PropTypes.number
     };
 
-    state = {
+    state: WPState = {
       width: 1280
     };
 
+    elementRef: ReactRef<HTMLDivElement> = React.createRef();
     mounted: boolean = false;
 
     componentDidMount() {
       this.mounted = true;
-
       window.addEventListener("resize", this.onWindowResize);
       // Call to properly set the breakpoint and resize the elements.
       // Note that if you're doing a full-width element, this can get a little wonky if a scrollbar
@@ -61,10 +77,12 @@ export default function WidthProvider<
 
     onWindowResize = () => {
       if (!this.mounted) return;
-      // eslint-disable-next-line react/no-find-dom-node
-      const node = ReactDOM.findDOMNode(this); // Flow casts this to Text | Element
-      if (node instanceof HTMLElement)
-        this.setState({ width: node.offsetWidth});
+      const node = this.elementRef.current; // Flow casts this to Text | Element
+      // fix: grid position error when node or parentNode display is none by window resize
+      // #924 #1084
+      if (node instanceof HTMLElement && node.offsetWidth) {
+        this.setState({ width: node.offsetWidth });
+      }
     };
 
     render() {
@@ -72,10 +90,22 @@ export default function WidthProvider<
       const { measureBeforeMount, ...rest } = this.props;
       if (measureBeforeMount && !this.mounted) {
         return (
-          <div className={this.props.className} style={this.props.style} />
+          <div
+            className={clsx(this.props.className, layoutClassName)}
+            style={this.props.style}
+            // $FlowIgnore ref types
+            ref={this.elementRef}
+          />
         );
       }
-      return <ComposedComponent {...rest} {...{'width': newWidth}} />;
+      return (
+        <ComposedComponent
+          innerRef={this.elementRef}
+          {...rest}
+          {...this.state}
+          {...{ 'width': newWidth }}
+        />
+      );
     }
   };
 }
